funken: add variadic transductions
This allows implementing zipping and other interesting operations in
terms of transduce.  Also, it is necessary for using transductions
as transformation for signals, since a signal might have multiple
parents.  For this API to make sense, the ranges are now passed as last
parameter, which is a bit against the std library conventions, which,
excepting std::transform(), lacks operations over multiple sequences.