Remove feature about fixing the type of filter
I am not convinced it is a useful feature anymore since anyway `reduce()`
fixes the type too and thus there must exist a conversion between both
branches.  It is easy to recover this feature if ever needed, but all
other tests pass without it so we leave it like this.

This fixes warning -Wconversion