variant: remove ambiguity when calling visitor
GCC 4.9 complained about this.  We were actually inheriting the call
operators of the actual visitors in a way that generates ambiguity
according to the standard.  It is surprising that clang did not complain
about this.  Since we can not use paremeter pack expansion in a using
declaration, we have to accumulate inheriting the different operators
one by one.