xform: make early termination really work
Before this change, early termination would be broken when a reducer
further calls `reduce`.  Now, a `reduce_nested` function is provided
that does not unwrap the potentially `reduced` value that is returned by
the reduction, such that an outer `reduce` call can stop early as well.

Also, this commit cleans up a bit the dispatching code that chooses the
right implementation of reduce for benchmarking purposes.