xform: make type erased transducers allow early termination
This change changes an important aspect of the whole transducer concept:
now transducers are not allowed to manipulate the state at all, nor wrap
it into a new type.  This is an aspect in which Clojure transducers are
inconsistent:

- They use the reducing function closure to hide the transducer internal
  state, instead of appending it to the accumulator, but...

- They denote the "reduced" state by wrapping the accumulator.

In my opinion, either (1) we always use the accumulator for state,
or (2) we never do, but an intermediate solution does not make sense.

In this commit we explire option 1.  Because the state type never
changes, our type erasure mechanism now always work.  Also, we can now
use exceptions to abort a reduction, and these can be type-erased as
well.

In later commits we will explore option 2.