xform: add type-erasing transducer<> type
This allows to erase the type of a transducer, such that different
transducers may be stored or passed around thorough places that can not
retain the full type.

There are some already acknowledged caveats already, that we shall
remove in the following commits:

- Can not type-erase transducers with variadic inputs.
- Early reductions are removed.
- The state type is fixed at both ends of the reduction.

Another caveat, not so easy to solve, is that the transducer is actually
applied on the first iteration in which it is used, not before.  This is
so because, until that moment, we don't really know what the type of the
state is.  Anyways, this should only be a problem for transducers that
have effects when they are applied over a reducer, and, dude, you should
really not be writing that kind of transducer :D

Note that though erasing the type of a transducer has a significant
performance cost (right now in the order of 70x).  This is because the
state needs to box the state in `boost::any` on every iteration.  There
are few ways we can improve this though by using a custom type erasing
box instead of `boost::any`.  Still, I doubt it can ever be lower than a
factor 5-10x so using `transducer<>` should be the exception, not the
norm.