xform: transducers are all purely functional
This commit changes how transducers that require `state` (like `take`)
work.  Unlike in Clojure, where the transducer hides its state in the
resulting reducer, here we use a purely functional approach.

When a transducer needs state, it can now append it to the accumulator
using `wrap_state` on the first iteration.  In the second iteration, it
can extract the data from the accumulator, compute, and return a new
wrapped state with updated data.  In this manner, it may manipulate extra
information during the reduction without carrying hidden state.

The API is designed in such a way that a single, generic, definition of
the reducer will suffice.  This is achieved by making `state_data`
return the initial extra state or the current state depending on whether
the passed in state is already wrapped or not.  This is needed for type
erasure to work.

The assumptions and type manipulations that a transducer may peform on
the accumulator type will be further formalized as concepts in later
commits.