funken: fix Clang does not like our use of declval
In case you are wondering why this `declval_` clone of `std::declval` is
needed at all... Clang is too cealous and it statica asserts too much
inside `declval` to make sure that you don't call it in an evaluated
context.

In particular, we use declval here inside function default arguments.
You are crazy Juanpe!  Don't worry, this function is never called, so it
is not a problem.  But for Clang it is, hence this clone.  Ideally I
would have never defined it, but then clang warns about a function with
internal linkage not being defined.