xform: fix type erased transducers coult not be type erased further
If a type erased transducer is composed and then type erased again,
errors would arise.  We need to follow the same technique that we used
to enable type erasure in other stateful transducers: implement them in
a way such that, while the resulting reducer might generate multiple
instantiations when dealing with concrete types, it will properly work
with *one single instantiation* over `any_state`.

In this case, we really want to deal with concrete types in the
implementation though, for performance mainly.  The `with_state`
function, that can dispatch both statically or dynamically to an
implementation of the transducer for a concrete unwrapped or wrapped
state type.

Also, for this new test to pass, `any_state::as<T>` now works properly
when `is_same<T, any_state>`.