xform: prevent allocation when type erasing simple transducers
When the any_state values are assigned, we reuse the memory that was
allocated for the last value that was held.  We make sure to chain such
assignments all the way through the reduction.

This optimization is not having effect yet for type erased stateful
reducers.

Also, this optimization is not exception safe.  We shall review this
later to minimize the effects of this.