meta: make concepts future proof
The latest Concepts-Lite papers [1][2] are quite unambiguous about the
fact that a `concept` *has to be* a function or constant -- it can't be
a struct that can converted to boolean, even if using it is
syntactically equivalent.  Thus our `concept<>` trick is not C++17
proof.

This commit makes sure that we define concepts in a future-proof way,
clearly separating the *concept specification* from the actual concept.
The concept is defined in such a way, that eventually we can just
redefine the `ABL_CONCEPT` macro to support real concepts.  A new
`ABL_CONCEPT_SPEC` macro is introduced to more easily define a concept
specification and it's accompanying concept.

This also makes naming a bit more consistent with Eric Nieblers
approach [3].  In particular:

  - We rename `requires` to `requires_` to avoid conflicts with the new
    C++17 keywords.

  - We rename `satisfies` to `models`...

The two approaches have still some disimiliraties.  In particular,
Eric's does not try to be C++17 proof and is more powerful -- i.e. it
allows to specify concept refinements.

[1] http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2013/n3701.pdf
[2] http://concepts.axiomatics.org/~ans/concepts-lite.pdf
[3] https://github.com/ericniebler/range-v3/blob/master/
      include/range/v3/utility/concepts.hpp