xform: make type erased transducers produce variadic outputs
There was a wrong assumption, which was embedded in the syntax of the
`transduce` type erasing type.  We assumed that a transducer can not
produce variadically, which is wrong.

The syntax has been corrected and now we make sure all transducers are
correct with further concept checks.  This has forced several fixes in
several concepts, which were not correctly modelled.