xform: fix `std::forward` after `std::forward` problem
We might have more of these, so we have to be careful and review the code.

It is hard to track, because in some cases it is legitimate, when we
know we are not sinking the argument, but extracting ortogonal parts of
it.  See `take` for an example of such.