testing: add micro-benchmarking framework
This makes it easier to write benchmarks that are "proper":

  - make multiple iterations, to make results significant
  - make multiple measurements, and take the minimum
  - can configure the amount of input

All these parameters can be customized at run-time so one can trade-off
accuracy and turn-around time, and explore how an algorithm evolves for
different input sizes.

This framework can be extended later, to include cool features like:

  - automatically measure for different input sizes.
  - assert that a certain algorithm is at max X-times slower than
    another one (useful for checking performance in CI).
  - check that all algorithms under test are actually giving the same
    results.

This commit also changes the naming conventions of the generated
binaries, to make them more unix-y.