xform: use new benchmarking framework to test transducers
There are minor improvements in the test cases themselves that make them
more accurate.  For example, unsigned types are used now to prevent UB
due to overflow if the data size is set to a too large number.  Also,
the stl-based "sum" copies the vector into the result in a potentially
more performant way.