xform: fix flattening of any_state values
The method `any_state.as<any_state>()` always succeeds, and
`any_state{any_state{}}` does not create a box for the `any_state`.
Thus `any_state` are flattened.  However this was not represented by the
result of `has<any_state>()`, which would always return `false`, and now
it always returns `true`.