xform: fix composition of non-determinstic and stateful transducers
We call **stateful transducer** one such that given a state of type `S`
and inputs of types `I...`, returns a type `S' = W<S, I...>`, that
depends on the current state and inputs.  Basically, it wraps the
current state of the reduction to attach its own state to
it (e.g. `take`).

A **non-deterministic transducer** is one that does not always call the
nested reducing function, depending the current state and inputs.

The problem with non-deterministic transducers is that, when composed
with a stateful transducer, they don't know what state type to return.
They do know the initial unwrapped state `S`, they do know the wrapped
state `S'`, but they don't know how to construct values of the
later. This is is because `S'` needs initialized by actual stateful
reducer with the initial attached state, for example, the total count
`n` for `take()`.

The solution is to return a union type `S'' = S + S'` that may contain
either or.  The `skip()` and `call()` functions take care of computing
what this `S''` type should be and wrapping and unwrapping the `S` and
`S'` values from it.  They also make try to avoid creating unnecessary
wrapper types, for example, `S'' = S` iff `S = S'`, and `S'' =
common_type<S, S'>` if such type exists (see `std::common_type` for
details [1])

[1] http://en.cppreference.com/w/cpp/types/common_type