xform: `reduce_nested` is non-deterministic
`reduce_nested` is like a non-deterministic transducer -- if the
sequence is empty and it is passed a stateful transducer, it should not
construct a state wrapper from an unwrapped state.  Instead, it should
just return a `skip_state` that may contain a wrapped state (if the
sequence is non-empty) or a raw state otherwise.

To highlight this problem at compile time `state_wrapper` such that is
not constructible from an unwrapped state.  Some new unit tests also
highlights the problem at run-time.

There are slight changes in the performance of `reduce` with the new
changes.  We have only tested on Clang.  We will do a new round of
optimizations later with GCC.