variant: use alternatives in variant to deduce return type
The type deduction algorithm before would just attempt to pass an
universally convertible value to every function passed in, and then
deduce the `common_type`... this would fail for generic functions and
other conversion edge-cases.

Instead, we now actually query what the result would be of destructuring
all the alternatives in the variant with the actual merged visitor, then
do the `common_type` of these.  This makes deduction always work as
expected and reduces the need for `when`, `otherwise`, etc...

This also changes:

- Testing of `boost::variant` and `eggs::variant` is unified, such that
  we make sure both implementations are feature complete.

- Specializing types in `eggs::variants::detail` namespace is no longer
  required to inherit from an `eggs::variant`.  We offer our own API to
  inherit from any variant type, which involves definining a nested
  `variant_types` or specializing `atria::variant::variant_types<>`.
  See the changes in `skip.hpp` for an example.