xform: add `replace` family of functions, a-la Clojure
These have small differences though.  The problem is that the Clojure
style one does not allow changing the type, something that might work
for Clojure's dynamic typing in some cases, but here it's more
convenient to have also something that replaces everything, thus
allowing to change the type.  The two versions added allow to either
default-construct missing elements or throw an exception.