xform: fix deducing complete type of type erased transducer
Before, the type erased transducer would fix the final type to whatever
type it received during the first iteration.  This prevents deducing the
result type of such transducer.

Deducing the output type of a transducer works by passing a reducing
function that just returns the inputs, and passing a dummy state that is
convertible to anything.  Type deduction will convert this dummy state
to the input state that is returned, and the input of the reducing
function (i.e. the output of the transducer) will get returned as the
actual state.

Now the type erased transducer properly queries the reducing function
for what state would it return, and then for what `state_complete` would
return, to see what the final type is.  This way the type wrapping chain
is not broken, allowing our type deduction trick to work.