travis: move to new docker based infrastructure
This allows the travis machines to start way faster.  This requires not
using `sudo`. The solution is to not install `libc++` in the system.
For some reason, when doing so, Clang does not detec anymore that it is
it's own code, so we need to disable some warnings with
`-Wno-reserved-id-macro`