xform: add `write` transducer to write to an output stream
For convenience, it can take a `terminator` to interperse between
inputs, and a `separator` to put between elements of a same input.

If we wanted to separate inputs with a form of separator instead of a
terminator, we would need to change the transducer to be stateful, so we
can distinguish the first iteration from the rest.  That is a plausible
improvement for the future.

Also, one may consider to change the implementation to use the
`interperse` or `interleave` transducers.  This is not possible because
these do not work heterogeneously.  An interesting improvement for the
future would be to change these to use `variant<>` when facing a
heterogeneous situation.  This would work with `write` out of the box
for `boost::variant<>`, since it supports stream operators [1]. Similar
operator would need to be manually defined for `eggs::variant` though.

[1] http://www.boost.org/doc/libs/1_59_0/boost/variant/detail/variant_io.hpp