xform: add `eager` transducer for holistic algorithms
There are certain algorithms that require the whole input sequence to be
able to do something.  Canonical examples are `sort` and `reverse`.
Even though transducers do not particulary suit these, sometimes it
might be interesting to express them as a transducer.  Most of the
actual transduction would then be delayed until the completion point, at
which the eager transducer can then proceed.