estd, meta: move `estd::Value_type` and add `detected` C++17 framework
So, I am starting to seriously doubt that a metafunction like the
`Value_type` that is often referenced in the Concepts Lite papers is
ever going to make it to the standard.  It doesn't belong in `estd::`.

However, what did make it to a C++ TS is the `detected` family of
functions for making the *voider* pattern easier to use.  This actually
simplifies the way one would implement such `Value_type` metafunction.

So this commit:

   - Removes `estd::Value_type`.

   - Adds `estd::detected` and friends.

   - Adds `meta::value_t` that effectively replaces the old
     `estd::Value_type` and improves its semantics.